Crate fabric_contract

This is the main crate for providing support for writing Smart Contracts to work with Hyperledger Fabric's Wasm chaincode runtime.

The fabric_contract_macros crate contains the macros to assist with development of the Smart Contracts

basic_contract_rs shows a simple Asset contract.

Modules

blockchain
Module to provide APIs to get information about Fabric

contract
Module to provide APIs to write contracts and interact with the ledger

data
Module to use to define the complex datatypes

prelude

runtime
Module to provide 'runtime' services.

Macros

register
Macro to use in the lib.rs file of your contract's crate

Attribute Macros

contract_impl
Use this macro to mark the implementation of the your contract structure

property
Define the properties of the datatype

transaction
Use this to mark the functions that are considered to be transaction functions

Derive Macros

DataTypeMacro
Use this to mark the structs that serve as complex data types Need to provide example