Module fabric_contract::contract

Module to provide APIs to write contracts and interact with the ledger

TODO: Split this into two modules?

Structs

Collection

Context

ContractError

ContractManager
Contract Manager structure that holds the list contract objects

Ledger
Ledger representing high level concept of the Fabric Blockchain network

State
A State is the combination of key and value that are contained within a collection.

Enums

CollectionName
Collection Name

Traits

Contract
Trait that is implemented for each contract Default implementations here

Routing

Attribute Macros

Contract_Impl
Use this macro to mark the implementation of the your contract structure

Transaction
Use this to mark the functions that are considered to be transaction functions