Struct fabric_contract::contract::State

pub struct State { /* fields omitted */ }

A State is the combination of key and value that are contained within a collection.

State-based endorsement should be set on this object as well.

Represents the things that are contained within the Collections.

Implementations

impl State

pub fn make_composite_key(keytype: String, attributes: Vec<String>) -> String

Make a composite key.

Remarks

Takes a set of strings and arranges them in a ':' separate single string

Arguments

keytype The first part of the key. Referred to as 'type' but is just the 1st string part of the key.

attributes Vector of Strings to make the rest of the ky

pub fn split_composite_key(key: String) -> Vec<String>

Splits a composite key

Splits the composite key into the parts, the first is the string that is passed in as the keytype above

pub fn new(key: String, data: Vec<u8>) -> State

Creates a new state

pub fn value(&self) -> Vec<u8>

Get the buffer that this state

pub fn key(&self) -> String

Get the key that used for this state

pub fn get_history(&self)

Returns an iterator of the state history for this state

pub fn get_hash(&self) -> Vec<u8>

gets the private hash for this stae

pub fn set_endorsment(&self)

Sets the State Based Endorsment for this state

pub fn get_endorsement(&self)